Added /gamemode, Removed /whitelist
